chore. Add files to the module
