+++ README.md
@@ -1,26 +1,27 @@
 # ![logo](https://raw.githubusercontent.com/azerothcore/azerothcore.github.io/master/images/logo-github.png) AzerothCore
+
 ## Starter Guild
-- Latest build status with azerothcore: [![Build Status](https://github.com/azerothcore/mod-starter-guild/workflows/core-build/badge.svg?branch=master&event=push)](https://github.com/azerothcore/mod-starter-guild)
+
+- Latest build status with azerothcore:
+
+[![Build Status](https://github.com/azerothcore/mod-starter-guild/workflows/core-build/badge.svg?branch=master&event=push)](https://github.com/azerothcore/mod-starter-guild)
 
 ## Description
 
 This module automatically assigns new characters to a specific guild depending on the faction.
 
-
 ## Usage
 
-- Create a new guild ingame for each faction with `.guild create` or directly into the database
-- Edit the configuration file
-- Create a new character and open the guild tab
-
+- Create a new guild ingame for each faction with `.guild create` or directly into the database.
+- Edit the configuration file.
+- Create a new character and open the guild tab.
 
 ## Credits
 
-* [stygiancore](  http://stygianthebest.github.io ): (Author of the module):
-* [SoulSeekkor]( https://github.com/SoulSeekkor ) cleaned up deprecation warnings and fixed naming
+* [stygiancore](  http://stygianthebest.github.io ): (Author of the module).
+* [SoulSeekkor]( https://github.com/SoulSeekkor ) cleaned up deprecation warnings and fixed naming.
 
 * AzerothCore: [repository](https://github.com/azerothcore) - [website](http://azerothcore.org/) - [discord chat community](https://discord.gg/PaqQRkd)
-
 
 ## License
